fix: resolve all 16 clippy warnings
- map_or(false, ...) → is_some_and(...) (5 instances across ngram, resolver, watch, parsers)
- &mut Vec<f64> → &mut [f64] (ptr_arg in calibrate/profile.rs)
- Manual Default impl → #[derive(Default)] (calibrate/resolver.rs)
- &"Legacy" → "Legacy" (op_ref in cli/clean.rs)
- loop { match rx.recv() } → while let Ok(events) = rx.recv() (cli/watch.rs)
- Nested if → collapsed if (detectors/mod.rs, detectors/unhandled_promise.rs)
- .map(|l| *l) → .copied() (detectors/missing_await.rs)
- let...else → ? operator (fixes/mod.rs)
- Remove unnecessary u32 as u32 cast (detectors/architectural_bottleneck.rs)
- vec![...] → slice literal (mcp/handlers.rs)

+++ repotoire-cli/src/detectors/architectural_bottleneck.rs
@@ -51,7 +51,7 @@ impl ArchitecturalBottleneckDetector {
         let high_complexity_threshold = config.get_option_or(
             "high_complexity_threshold",
             config.adaptive.warn_usize(MetricKind::Complexity, 20) as u32,
-        ) as u32;
+        );
         let min_fan_in = config.get_option_or(
             "min_fan_in",
             config.adaptive.warn_usize(MetricKind::FanIn, 15),
